refactored a couple of components, show escalation chain
