+++ grafana-plugin/src/pages/integration_2/Integration2.module.scss
@@ -81,15 +81,6 @@ $ITEMS-MARGIN: 24px;
   padding: 15px;
 }
 
-.input {
-  &--short {
-    width: 500px;
-  }
-  &--long {
-    width: 700px;
-  }
-}
-
 .routesSection {
   padding-top: 12px;
 }
