+++ grafana-plugin/src/pages/integration_2/IntegrationRouteDisplay.module.scss
@@ -1,0 +1,8 @@
+.input {
+  &--short {
+    width: 500px;
+  }
+  &--long {
+    width: 700px;
+  }
+}
